feat(codex): wire the collab card engine into the live delegation path (task 16)
Rewrites thread_item_render.rs's CollabAgentToolCall/SubAgentActivity/
AgentMessage arms and event_mapper.rs's item/started collab dispatch to call
collab_card.rs instead of the retired collab_render.rs, which is fully
superseded and deleted. Splits collab_card.rs's resolution-side helpers
(resolve_card, on_wait_started/completed, fail_all_open_cards) into a
collab_resolve.rs sibling to stay under the 300-line ceiling.

event_mapper.rs's handle_item_completed keeps its existing card_for_thread
ParentIdSink wrapping and handle_item_started/handle_turn_* keep their
current shape — the Owner-enum routing rewrite is task 17/18's scope.
